requirements.txt
langchain
chromadb
ollama
gitpython
typer
rich
tqdm
pytest
